requirements.txt
absl-py==1.0.0
adal==1.2.7
ajax==3.0.0
argcomplete==1.12.3
asgiref==3.4.1
astunparse==1.6.3
azure-common==1.1.27
azure-core==1.20.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==19.0.0
azure-mgmt-storage==19.0.0
azureml-core==1.37.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
cryptography==3.4.8
decorator==5.1.0
Django==4.0
django-appconf==1.0.5
docker==5.0.3
efficientnet==1.0.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.4.0
glob2==0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==3.6.0
humanfriendly==9.2
idna==3.3
image-classifiers==1.0.0
imageio==2.13.3
importlib-metadata==4.8.2
isodate==0.6.1
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpickle==2.0.0
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
knack==0.8.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
paramiko==2.8.1
pathspec==0.9.0
Pillow==8.4.0
pkginfo==1.8.2
portalocker==2.3.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
SecretStorage==3.3.1
segmentation-models==1.0.1
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tifffile==2021.11.2
typing_extensions==4.0.1
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
